* lib/Makefile.am
* lib/lp-common.c
* lib/lp-media.c
* lib/lp-scene.c
* lib/play-internal.h
* src/appsrctext.c
* tests/Makefile.am
* tests/lua.c
* tests/test-lp-media-new.c
* tests/test-lp-scene-advance.c
* tests/test-lp-scene-prop-pattern.c -> tests/test-lp-scene-prop-background.c
* tests/test-lp-scene-prop-interval.c
* tests/test-lua-play0-scene-get.lua
* tests/test-lua-play0-scene-receive.lua
* tests/test-lua-play0-scene-set.lua
* tests/tests.h

+++ tests/lua.c
@@ -44,7 +44,7 @@ main (int argc, const char **argv)
   for (i = 0; i < argc; i++)
     {
       lua_pushstring (L, argv[i]);
-      lua_rawseti (L, -2, i);
+      lua_rawseti (L, -2, (int) i);
     }
   lua_setglobal (L, "arg");
